PM / devfreq: memlat: Improve tracing of stats and votes
When the memlat governor votes for 0, it didn't generate any "update" trace
points to avoid spamming the trace buffer. But this can sometimes make it
harder to interpret some traces logs.

So, as a middle ground betweem making the trace harder to interpret vs
spamming the trace buffer, generate one trace point, when the vote
transitions from a non-zero value to a zero.

At the same time, the memlat governor was also generating "meas" trace
points for completely idle cores during every polling interval. This causes
a significant amount of spamming of the trace buffer. So, when a core is
idle, don't generate any "meas" trace points for it.

Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Change-Id: I3c72cde87bf56e45e31cc04fc4f80c030eed1e1a
Signed-off-by: celtare21 <celtare21@gmail.com>